% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat2ometiff.R
\name{.InitalizeSparseMatrix}
\alias{.InitalizeSparseMatrix}
\title{InitalizeSparseMatrix}
\usage{
.InitalizeSparseMatrix(x_min, x_max, y_min, y_max)
}
\arguments{
\item{x_min}{The minimum x-coordinate of the global image.}

\item{x_max}{The maximum x-coordinate of the global image.}

\item{y_min}{The minimum y-coordinate of the global image.}

\item{y_max}{The maximum y-coordinate of the global image.}
}
\value{
A sparse matrix with the dimensions of the CosMx image.
}
\description{
This function initializes a sparse matrix with the dimensions of the global image.
}
