% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat2ometiff.R
\name{.ReadFlatFiles}
\alias{.ReadFlatFiles}
\title{ReadFlatFiles}
\usage{
.ReadFlatFiles(expressionCsvPath = NULL, polygonCsvPath = NULL)
}
\arguments{
\item{expressionCsvPath}{Path to the expression matrix CSV file.}

\item{polygonCsvPath}{Path to the polygon CSV file.}
}
\value{
A list containing the expression and polygon data tables.
}
\description{
This function reads in the flat files from CosMx.
}
