% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat2ometiff.R
\name{Flat2Matrix}
\alias{Flat2Matrix}
\title{Flat2Matrix}
\usage{
Flat2Matrix(
  expressionCsvPath = NULL,
  polygonCsvPath = NULL,
  basePath = "./sparse_matrix_files/",
  channelNamesAreExpressionCSVColumnNames = TRUE
)
}
\arguments{
\item{echo}{boolean to control responses printed (or is it???)}
}
\description{
Temporary function to set up R package
}
