% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flat2OMETiff-package.R
\docType{package}
\name{Flat2OMETiff-package}
\alias{Flat2OMETiff}
\alias{Flat2OMETiff-package}
\title{Flat2OMETiff: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: GW McElfresh \email{mcelfreshgw@gmail.com} (\href{https://orcid.org/0000-0002-1948-7571}{ORCID})

}
\keyword{internal}
