% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat2ometiff.R
\name{StitchTiffChannels}
\alias{StitchTiffChannels}
\title{StitchTiffChannels}
\usage{
StitchTiffChannels(
  tiffChannelDirectory = "./tiff_channels/",
  outputDirectory = "./stitched_tiff/",
  channelWhitelist = NULL
)
}
\arguments{
\item{tiffChannelDirectory}{Path to the directory containing the tiff channels.}

\item{outputDirectory}{Path to the directory where the stitched tiff files will be written.}

\item{channelWhitelist}{A vector of channel names to include in the stitched tiff files.}
}
\description{
This function stitches the tiff channels together using the python tifffile library.
}
