% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat2ometiff.R
\name{Flat2Matrix}
\alias{Flat2Matrix}
\title{Flat2Matrix}
\usage{
Flat2Matrix(
  expressionCsvPath = NULL,
  polygonCsvPath = NULL,
  outputDirectory = "./sparse_matrix_files/",
  channelNamesAreExpressionCSVColumnNames = TRUE
)
}
\arguments{
\item{expressionCsvPath}{Path to the expression matrix CSV file.}

\item{polygonCsvPath}{Path to the polygon CSV file.}

\item{outputDirectory}{Path to the directory where the sparse matrices will be written.}

\item{channelNamesAreExpressionCSVColumnNames}{Boolean to control whether the channel names are taken from the expression CSV file column names. If not, the channels are indexed numerically.}
}
\description{
This function converts flat files from CosMx into a directory of sparse matrices. One file for each channel will be created within the \code{outputDirectory} directory.
}
